Make TimelineEvent.end_date a nullable field
The field is allowed to be blank, and thus Django does not require
that it be filled in; however, in the database, is it _not_ a nullable
field, so the database _does_ require that a value be set. This
difference in handling between Django and the database engine creates
an IntegrityError when events are saved without end dates. This commit
removes that problem.

Signed-off-by: Michael Dippery <mdippery@nybooks.com>