Handle timeline PKs that are longs
If a `src` is not specified when calling the `timeline` tag, `timeline`
takes the pk of the variable `timeline` in the current context. This
is generally a long (not an int), but the tag was only handling ints.
This caused an exception in the fallback case, in which it was assumed
that `src` had an `isdigit()` method (probably assuming that `src`
would be a `str` if it was not an `int`).

This commit handles cases in which `src` is a long (perhaps the most
common case), and falls back on converting a string to a long when
necessary.

Signed-off-by: Michael Dippery <mdippery@nybooks.com>